# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.CantorListResponse:
    properties:
      displayName:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  main.RatesResponse:
    properties:
      buyRate:
        type: string
      cantorID:
        type: integer
      currency:
        type: string
      sellRate:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the backend API for the Gix exchange rate application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gix API
  version: "1.0"
paths:
  /cantors:
    get:
      description: Returns a list of all available cantors with their geolocations.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.CantorListResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List Cantors
      tags:
      - cantors
  /healthz:
    get:
      description: Checks if the server, database, and Redis are running correctly.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health Check
      tags:
      - health
  /rates:
    get:
      description: Returns the buy and sell rates for a specific cantor and currency.
        Scrapes data in real-time if not cached.
      parameters:
      - description: Cantor ID
        in: query
        name: cantor_id
        required: true
        type: integer
      - description: Currency Code (e.g., EUR, USD)
        in: query
        name: currency
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RatesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Rates
      tags:
      - rates
swagger: "2.0"